[1.0.17] CONDITION: developing  - Added TraderJoeMath.py, TraderJoeSwap.py  - Introduced Bases  - Now Bin and Tick are global union

+++ src/DataFetcher/api_clients/Raydium/RaydiumMetadataScheme.py
@@ -1,4 +1,4 @@
 from src.DataFetcher.api_clients.BaseMetadataScheme import BaseMetadataScheme
 
 class RaydiumMetadataScheme(BaseMetadataScheme):
-    feeRate: int
+    feeRate: float
